fix: moved LibraryRetrieveSourcePathExecutor into its own file
